User auth added
