+++ Threads/App/ThreadsApp.swift
@@ -24,7 +24,7 @@ struct ThreadsApp: App {
     @UIApplicationDelegateAdaptor(AppDelegate.self) var delegate
     var body: some Scene {
         WindowGroup {
-            LoginView()
+            ContentView()
         }
     }
 }
