--- Threads/Core/Root/ContentView.swift
@@ -1,26 +1,0 @@
-//
-//  ContentView.swift
-//  Threads
-//
-//  Created by aykut ipek on 12.09.2023.
-//
-
-import SwiftUI
-
-struct ContentView: View {
-    var body: some View {
-        VStack {
-            Image(systemName: "globe")
-                .imageScale(.large)
-                .foregroundColor(.accentColor)
-            Text("Hello, world!")
-        }
-        .padding()
-    }
-}
-
-struct ContentView_Previews: PreviewProvider {
-    static var previews: some View {
-        ContentView()
-    }
-}
